Move properties/funcs into TISubNodeMixin (#644)
Small PR on top of a WIP to avoid making suggestions/commits

---------
